WIP: QueryExecutor now as I want it. Its typescript API describes the type or structure of the parameters.
WhereClause in collection.js will generate ExpressionCollections that have its data set to a structure of {keyPath, value}, {keyPath, values} or {lExpr, rExpr}

Now, need to check that collection.js is a complete implementation.

Need to define how algorithms fit into this framework. Specifically we need to define how an algorithm can affect the IDBCursor to continue to a certain key.

NOTE: dexie lab code "idbCollection" needs to be updated to instead extend the QueryExector class.